backend/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
requests>=2.31.0
pandas>=2.0.0
numpy>=1.26.0
python-dotenv>=1.0.0
polygon-api-client>=1.13.0
gunicorn>=21.0.0
